[Fix #251] showModal Socket Event 수정
